01_Publishing/generator/requirements.txt
pandas==1.5.3
protobuf==4.21.12
faker
google.auth
google.cloud.pubsub
